+++ app/lib/pages/debug/debug_page.dart
@@ -6,6 +6,7 @@ import 'package:localsend_app/pages/debug/discovery_debug_page.dart';
 import 'package:localsend_app/pages/debug/http_logs_page.dart';
 import 'package:localsend_app/pages/debug/security_debug_page.dart';
 import 'package:localsend_app/provider/app_arguments_provider.dart';
+import 'package:localsend_app/provider/persistence_provider.dart';
 import 'package:localsend_app/widget/debug_entry.dart';
 import 'package:refena_flutter/refena_flutter.dart';
 import 'package:routerino/routerino.dart';
@@ -59,6 +60,10 @@ class DebugPage extends StatelessWidget {
                   onPressed: () async => context.push(() => const RefenaTracingPage()),
                   child: const Text('Refena Tracing'),
                 ),
+              FilledButton(
+                onPressed: () async => await context.ref.read(persistenceProvider).clear(),
+                child: const Text('Clear settings'),
+              ),
             ],
           ),
         ],