+++ app/lib/provider/network/nearby_devices_provider.dart
@@ -1,15 +1,11 @@
 import 'dart:async';
 
-import 'package:dio/dio.dart';
-import 'package:localsend_app/constants.dart';
 import 'package:localsend_app/model/device.dart';
-import 'package:localsend_app/model/dto/info_dto.dart';
+import 'package:localsend_app/model/persistence/favorite_device.dart';
 import 'package:localsend_app/model/state/nearby_devices_state.dart';
-import 'package:localsend_app/provider/dio_provider.dart';
 import 'package:localsend_app/provider/logging/discovery_logs_provider.dart';
 import 'package:localsend_app/provider/network/multicast_provider.dart';
-import 'package:localsend_app/provider/security_provider.dart';
-import 'package:localsend_app/util/api_route_builder.dart';
+import 'package:localsend_app/provider/network/targeted_discovery_provider.dart';
 import 'package:localsend_app/util/task_runner.dart';
 import 'package:logging/logging.dart';
 import 'package:refena_flutter/refena_flutter.dart';
@@ -22,26 +18,29 @@ final _logger = Logger('NearbyDevices');
 ///
 /// Use [scanProvider] to have a high-level API to perform discovery operations.
 final nearbyDevicesProvider = NotifierProvider<NearbyDevicesNotifier, NearbyDevicesState>((ref) {
-  return NearbyDevicesNotifier();
+  return NearbyDevicesNotifier(
+    targetedDiscoveryService: ref.read(targetedDiscoveryProvider),
+    multicastService: ref.read(multicastProvider),
+  );
 });
 
 Map<String, TaskRunner> _runners = {};
 
 class NearbyDevicesNotifier extends Notifier<NearbyDevicesState> {
-  late Dio _dio;
-  late String _fingerprint;
-  late MulticastService _multicastService;
+  final TargetedDiscoveryService _targetedDiscoveryService;
+  final MulticastService _multicastService;
+
+  NearbyDevicesNotifier({
+    required TargetedDiscoveryService targetedDiscoveryService,
+    required MulticastService multicastService,
+  })  : _targetedDiscoveryService = targetedDiscoveryService,
+        _multicastService = multicastService;
 
   @override
-  NearbyDevicesState init() {
-    _dio = ref.read(dioProvider).discovery;
-    _fingerprint = ref.read(securityProvider).certificateHash;
-    _multicastService = ref.read(multicastProvider);
-    return const NearbyDevicesState(
-      runningIps: {},
-      devices: {},
-    );
-  }
+  NearbyDevicesState init() => const NearbyDevicesState(
+        runningIps: {},
+        devices: {},
+      );
 
   /// Binds the UDP port and listens for incoming announcements.
   /// This should run forever as long as the app is running.
@@ -60,7 +59,7 @@ class NearbyDevicesNotifier extends Notifier<NearbyDevicesState> {
 
   /// Scans one particular subnet with traditional HTTP/TCP discovery.
   /// This method awaits until the scan is finished.
-  Future<void> startScan({required int port, required String localIp, required bool https}) async {
+  Future<void> startLegacyScan({required int port, required String localIp, required bool https}) async {
     if (state.runningIps.contains(localIp)) {
       // already running for the same localIp
       return;
@@ -68,25 +67,49 @@ class NearbyDevicesNotifier extends Notifier<NearbyDevicesState> {
 
     state = state.copyWith(runningIps: {...state.runningIps, localIp});
 
-    await for (final device in _getStream(localIp, port, https, _fingerprint)) {
+    await for (final device in _getStream(localIp, port, https)) {
       registerDevice(device);
     }
 
     state = state.copyWith(runningIps: state.runningIps.where((ip) => ip != localIp).toSet());
   }
 
-  Stream<Device> _getStream(String networkInterface, int port, bool https, String fingerprint) {
+  Stream<Device> _getStream(String networkInterface, int port, bool https) {
     final ipList = List.generate(256, (i) => '${networkInterface.split('.').take(3).join('.')}.$i').where((ip) => ip != networkInterface).toList();
     _runners[networkInterface]?.stop();
     _runners[networkInterface] = TaskRunner<Device?>(
       initialTasks: List.generate(
         ipList.length,
-        (index) => () async => _doRequest(ipList[index], port, https, fingerprint),
+        (index) => () async => _doRequest(ipList[index], port, https),
       ),
       concurrency: 50,
     );
 
     return _runners[networkInterface]!.stream.where((device) => device != null).cast<Device>();
+  }
+
+  Future<void> startFavoriteScan({required List<FavoriteDevice> devices, required bool https}) async {
+    if (devices.isEmpty) {
+      return;
+    }
+    await for (final device in _getFavoriteStream(devices: devices, https: https)) {
+      registerDevice(device);
+    }
+  }
+
+  Stream<Device> _getFavoriteStream({required List<FavoriteDevice> devices, required bool https}) {
+    final runner = TaskRunner<Device?>(
+      initialTasks: List.generate(
+        devices.length,
+        (index) => () async {
+          final device = devices[index];
+          return _doRequest(device.ip, device.port, https);
+        },
+      ),
+      concurrency: 50,
+    );
+
+    return runner.stream.where((device) => device != null).cast<Device>();
   }
 
   void registerDevice(Device device) {
@@ -99,23 +122,18 @@ class NearbyDevicesNotifier extends Notifier<NearbyDevicesState> {
     state = state.copyWith(devices: {});
   }
 
-  Future<Device?> _doRequest(String currentIp, int port, bool https, String fingerprint) async {
+  Future<Device?> _doRequest(String currentIp, int port, bool https) async {
     _logger.fine('Requesting $currentIp');
-    // We use the legacy route to make it less breaking for older versions
-    final url = ApiRoute.info.targetRaw(currentIp, port, https, peerProtocolVersion);
-    Device? device;
-    try {
-      final response = await _dio.get(url, queryParameters: {
-        'fingerprint': fingerprint,
-      });
-      final dto = InfoDto.fromJson(response.data);
-      device = dto.toDevice(currentIp, port, https);
+    final device = await _targetedDiscoveryService.discover(
+      ip: currentIp,
+      port: port,
+      https: https,
+      onError: null,
+    );
+    if (device != null) {
       ref.notifier(discoveryLogsProvider).addLog('[DISCOVER/TCP] ${device.alias} (${device.ip}, model: ${device.deviceModel})');
-    } on DioException catch (_) {
-      device = null;
-    } catch (e) {
-      device = null;
     }
+
     return device;
   }
 }