+++ app/assets/CHANGELOG.md
@@ -1,10 +1,12 @@
 ## 1.12.0 (2023-)
 
+- feat: add favorites (@Tienisto)
 - feat: drastically improve GPU usage by optimizing the spin animation (@Tienisto)
 - feat: add OLED color mode (@Tienisto)
 - feat: use better colors for the toggle switches in the settings (@gitstart)
 - feat(desktop): support pasting from clipboard (@gitstart, @Tienisto)
 - feat(linux): allow disabling client side decorations on Wayland (@I-Want-ToBelieve)
+- feat(android): use high framerate on devices that lock at 60 Hz like on some OnePlus phones (@Tienisto)
 - fix(desktop): fallback to "$HOME/Downloads" when default downloads folder is unavailable (@Sqbika)
 
 ## 1.11.1 (2023-09-04)